# Conflicts:
#	terrain/block_thread_manager.h
#	terrain/voxel_terrain.cpp

+++ util/utility.h
@@ -11,10 +11,10 @@
 // Takes elements starting from a given position and moves them at the beginning,
 // then shrink the array to fit them. Other elements are discarded.
 template <typename T>
-void shift_up(Vector<T> &v, int pos) {
+void shift_up(Vector<T> &v, unsigned int pos) {
 
-	int j = 0;
-	for (int i = pos; i < v.size(); ++i, ++j) {
+	unsigned int j = 0;
+	for (unsigned int i = pos; i < (unsigned int)v.size(); ++i, ++j) {
 		v.write[j] = v[i];
 	}
 
@@ -23,10 +23,10 @@ void shift_up(Vector<T> &v, int pos) {
 }
 
 template <typename T>
-void shift_up(std::vector<T> &v, int pos) {
+void shift_up(std::vector<T> &v, unsigned int pos) {
 
-	int j = 0;
-	for (int i = pos; i < v.size(); ++i, ++j) {
+	unsigned int j = 0;
+	for (unsigned int i = pos; i < v.size(); ++i, ++j) {
 		v[j] = v[i];
 	}
 
@@ -37,7 +37,7 @@ void shift_up(std::vector<T> &v, int pos) {
 // Pops the last element of the vector and place it at the given position.
 // (The element that was at this position is the one removed).
 template <typename T>
-void unordered_remove(Vector<T> &v, int pos) {
+void unordered_remove(Vector<T> &v, unsigned int pos) {
 	int last = v.size() - 1;
 	v.write[pos] = v[last];
 	v.resize(last);
